+++ examples/array_simple.rb
@@ -12,9 +12,7 @@ report = Benchmark.bigo do |x|
   x.generator {|size| (0...size).to_a.shuffle }
 
   # specifies how the size of the object should grow
-  #   options: linear, exponential
-  #
-  # specifies that arrays should grow linearly by 1000
+  # indicates that arrays should grow linearly by 1000
   x.linear 1000
 
   # report takes a label and a block.